requirements.txt
awswrangler==3.12.1
boto3==1.39.3
botocore==1.39.3
clickhouse-connect==0.7.0
colorama==0.4.6
jmespath==1.0.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
